requirements.txt
# Flask Backend for Academic ERP
Flask==3.0.0
Werkzeug==3.0.1

# Database
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9

# Authentication
PyJWT==2.8.0
cryptography==41.0.7

# API & CORS
flask-cors==4.0.0

# Environment & Configuration
python-dotenv==1.0.0

# Production Server
gunicorn==21.2.0

# AI Integration
google-generativeai==0.3.1

# Documentation
flasgger==0.9.7.1
